+++ CCIA.SQL/dbo/Tables/countries.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[countries] (
-    [country_id]    SMALLINT      IDENTITY (1, 1) NOT NULL,
+    [country_id]    INT           IDENTITY (1, 1) NOT NULL,
     [country_code]  VARCHAR (3)   NOT NULL,
     [country_name]  VARCHAR (100) NOT NULL,
     [oecd_member]   BIT           CONSTRAINT [DF_countries_oecd] DEFAULT ((0)) NOT NULL,
@@ -7,4 +7,6 @@
     [user_modified] VARCHAR (9)   NULL,
     CONSTRAINT [PK_Countries] PRIMARY KEY CLUSTERED ([country_id] ASC)
 );
+
+
